+++ cfdm/datum.py
@@ -26,27 +26,27 @@ class Datum(mixin.Parameters, mixin.NetCDFVariable, mixin.Files, core.Datum):
 
     """
 
-    def __init__(self, parameters=None, source=None, copy=True):
-        """**Initialisation**
-
-        :Parameters:
-
-            parameters: `dict`, optional
-               Set parameters. The dictionary keys are parameter
-               names, with corresponding values.
-
-               Parameters may also be set after initialisation with
-               the `set_parameters` and `set_parameter` methods.
-
-               *Parameter example:*
-                 ``parameters={'earth_radius': 6371007.}``
-
-            {{init source: optional}}
-
-            {{init copy: `bool`, optional}}
-
-        """
-        super().__init__(parameters=parameters, source=source, copy=copy)
-
-        self._initialise_netcdf(source)
-        self._initialise_original_filenames(source)
+    #def __init__(self, parameters=None, source=None, copy=True):
+    #    """**Initialisation**
+    #
+    #    :Parameters:
+    #
+    #        parameters: `dict`, optional
+    #           Set parameters. The dictionary keys are parameter
+    #           names, with corresponding values.
+    #
+    #           Parameters may also be set after initialisation with
+    #           the `set_parameters` and `set_parameter` methods.
+    #
+    #           *Parameter example:*
+    #             ``parameters={'earth_radius': 6371007.}``
+    #
+    #        {{init source: optional}}
+    #
+    #        {{init copy: `bool`, optional}}
+    #
+    #    """
+    #    super().__init__(parameters=parameters, source=source, copy=copy)
+    #
+    #    self._initialise_netcdf(source)
+    #    self._initialise_original_filenames(source)