+++ cfdm/domain.py
@@ -112,8 +112,8 @@ class Domain(
                 if mesh_id is not None:
                     self.set_mesh_id(mesh_id)
 
-        self._initialise_netcdf(source)
-        self._initialise_original_filenames(source)
+        #self._initialise_netcdf(source)
+        #self._initialise_original_filenames(source)
 
         self._set_dataset_compliance(self.dataset_compliance(), copy=True)