+++ cfdm/field.py
@@ -33,7 +33,7 @@ logger = logging.getLogger(__name__)
 
 
 class Field(
-    mixin.QuantizationWriteMixin,
+    mixin.QuantizationMixin,
     mixin.NetCDFVariable,
     mixin.NetCDFGeometry,
     mixin.NetCDFGlobalAttributes,
@@ -157,10 +157,8 @@ class Field(
                 if mesh_id is not None:
                     self.set_mesh_id(mesh_id)
 
-            self._init_quantization(source, copy)
-
-        self._initialise_netcdf(source)
-        self._initialise_original_filenames(source)
+#        self._initialise_netcdf(source)
+#        self._initialise_original_filenames(source)
 
         self._set_dataset_compliance(self.dataset_compliance(), copy=True)