+++ cfdm/mixin/quantizationmixin.py
@@ -1,9 +1,108 @@
-class QuantizationWriteMixin:
-    """Mixin class for constructs that allow quantize-on-write.
+class QuantizationMixin:
+    """Mixin class for access to quantization metdata.
 
     .. versionadded:: (cfdm) NEXTVERSION
 
     """
+
+    def __initialise(self, source, copy=True):
+        """Initialise quantization information from a source.
+
+        Intended to be called from `__init__` via
+        `_{{class}}__initialise`
+
+        .. versionadded:: (cfdm) NEXTVERSION
+
+        :Returns:
+
+            `None`
+
+        """
+        try:
+            q = source.get_quantization(None)
+        except (AttributeError, ValueError, TypeError):
+            pass
+        else:
+            if q is not None:
+                self._set_quantization(q, copy=copy)
+
+        try:
+            q = source.get_quantize_on_write(None)
+        except (AttributeError, ValueError, TypeError):
+            pass
+        else:
+            if q is not None:
+                self.set_quantize_on_write(q)
+
+    def _del_quantization(self, default=ValueError()):
+        """Remove quantization metadata.
+
+        .. warning:: Removing quantization metadata does not change
+                     the actual quantization (if any) of the data.
+
+        .. versionadded:: (cfdm) NEXTVERSION
+
+        .. seealso:: `get_quantization`, `_set_quantization`,
+                     `del_quantize_on_write`
+
+        :Parameters:
+
+            default: optional
+                Return the value of the *default* parameter if there
+                is no quantization metadata.
+
+                {{default Exception}}
+
+        :Returns:
+
+            `Quantization`
+                The deleted quantization metadata.
+
+        """
+        return self._del_component("quantization", default)
+
+    def _set_quantization(self, quantization, copy=True):
+        """Set quantization metadata.
+
+        Any existing quantization metadata removed prior to the new
+        setting.
+
+        .. warning:: Setting quantization metadata does not change the
+                     actual quantization (if any) of the data, and
+                     will not cause the data to be quantized when
+                     written to a netCDF dataset.
+
+        .. versionadded:: (cfdm) NEXTVERSION
+
+        .. seealso:: `get_quantization`, `_del_quantization`,
+                     `set_quantize_on_write`
+
+        :Parameters:
+
+            quantization: `Quantization`
+                The new quantization metadata.
+
+            copy: `bool`, optional
+                If True (the default) then copy *quantization* prior
+                to insertion.
+
+        :Returns:
+
+            `None`
+
+        """
+        q_on_w = self.get_quantize_on_write(None)
+        if q_on_w is not None:
+            raise ValueError(
+                "Can't set quantization metadata on a "
+                f"{self.__class__.__name__} that has a quantize-on-write "
+                f"instruction: {q_on_w!r}"
+            )
+
+        if copy:
+            quantization = quantization.copy()
+
+        return self._set_component("quantization", quantization, copy=False)
 
     def del_quantize_on_write(self, default=ValueError()):
         """Remove a quantize-on-write instruction.
@@ -12,8 +111,8 @@ class QuantizationWriteMixin:
 
         .. versionadded:: (cfdm) NEXTVERSION
 
-        .. seealso:: `get_quantization`, `get_quantize_on_write`,
-                     `set_quantize_on_write`
+        .. seealso:: `get_quantize_on_write`, `set_quantize_on_write`,
+                     `_del_quantization`
 
         :Parameters:
 
@@ -32,6 +131,26 @@ class QuantizationWriteMixin:
         """
         return self._del_component("quantize_on_write", default)
 
+    def get_quantization(self, default=ValueError()):
+        """Get quantization metadata.
+
+        .. versionadded:: (cfdm) NEXTVERSION
+
+        .. seealso:: `_del_quantization`, `_set_quantization`,
+                     `get_quantize_on_write`
+
+        :Returns:
+
+            `Quantization`
+                A copy of quantization metadata.
+
+        """
+        q = self._get_component("quantization", None)
+        if q is None:
+            return self._default(default)
+
+        return q.copy()
+
     def get_quantize_on_write(self, default=ValueError()):
         """Get a quantize-on-write instruction.
 
@@ -39,8 +158,8 @@ class QuantizationWriteMixin:
 
         .. versionadded:: (cfdm) NEXTVERSION
 
-        .. seealso:: `get_quantization`, `set_quantize_on_write`,
-                     `del_quantize_on_write`
+        .. seealso:: `set_quantize_on_write`, `del_quantize_on_write`,
+                     `get_quantization`
 
         :Parameters:
 
@@ -53,11 +172,14 @@ class QuantizationWriteMixin:
         :Returns:
 
             `Quantization`
-                The quantize-on-write instruction in a quantization
-                variable.
+                A copy of the quantize-on-write instruction.
 
         """
-        return self._get_component("quantize_on_write", default)
+        q = self._get_component("quantize_on_write", None)
+        if q is None:
+            return self._default(default)
+
+        return q.copy()
 
     def set_quantize_on_write(
         self,
@@ -68,16 +190,17 @@ class QuantizationWriteMixin:
     ):
         """Set a quantize-on-write instruction.
 
-        Calling `set_quantize_on_write` does not immediataely change
-        the data, but if the {{class}} is written to a netCDF dataset
-        with `{{package}}.write` then the quantization will be applied
-        to its data on disk, leaving the construct's data in memory
-        unchanged.
+        .. note:: Setting a quantize-on-write instruction does not
+                  immediately change the data, but if the `{{class}}`
+                  is written to a netCDF dataset with
+                  `{{package}}.write`, then the quantization will be
+                  applied at that time to the data on disk, leaving
+                  the construct's data in memory unchanged.
 
         .. versionadded:: (cfdm) NEXTVERSION
 
-        .. seealso:: `get_quantization`, `get_quantize_on_write`,
-                     `del_quantize_on_write`
+        .. seealso:: `get_quantize_on_write`, `del_quantize_on_write`,
+                     `_set_quantization`
 
         :Parameters:
 
@@ -119,18 +242,20 @@ class QuantizationWriteMixin:
         overriding its per-variable quantization parameter:
 
         >>> q = g.get_quantization()
-        >>> f.set_quantize_on_write(q, 'quantization_nsd': 6}))
+        >>> f.set_quantize_on_write(q, quantization_nsd=13))
 
         """
-        if self.get_quantization(None) is not None:
+        qm = self.get_quantization(None)
+        if qm is not None:
             raise ValueError(
                 "Can't set a quantize-on-write instruction on a "
-                f"{self.__class__.__name__} that is already quantized"
+                f"{self.__class__.__name__} that has quantization metadata: "
+                f"{qm!r}"
             )
 
         if quantization is None:
-            # Note: A parent class must define `_Quantization` as its
-            #       quantization component.
+            # Note: A parent class must define the `_Quantization`
+            #       attribute as its quantization component class.
             q = self._Quantization()
         else:
             q = quantization.copy()
@@ -142,37 +267,37 @@ class QuantizationWriteMixin:
         else:
             algorithm = q.get_parameter("algorithm", None)
 
+        # Get the parameter name associated with this algorithm
         parameter = q.algorithm_parameters().get(algorithm)
-
         if parameter is None:
             raise ValueError(
-                "Must provide a quantization algorthm with the "
-                "'quantization' or 'algorithm' keywords "
+                "Invalid quantization algorithm: {algorithm!r}. "
+                f"Must be one of {tuple(q.algorithm_parameters())}"
             )
 
         if quantization_nsd is not None:
             # Override the quantization_nsd parameter
             if quantization_nsb is not None:
                 raise ValueError(
-                    "Can not set both keywords 'quantization_nsd' and "
+                    "Can't set keywords 'quantization_nsd' and "
                     "'quantization_nsb' at the same time"
                 )
 
             if parameter != "quantization_nsd":
                 raise ValueError(
-                    f"Can only provide keyword {parameter!r} for "
+                    f"Can set keyword {parameter!r} for "
                     f"quantization algorithm {algorithm!r}"
                 )
 
-            q.set_parameter(parameter, quantization_nsd, copy=False)
+            q.set_parameter(parameter, quantization_nsd)
         elif quantization_nsb is not None:
             # Override the quantization_nsb parameter
             if parameter != "quantization_nsb":
                 raise ValueError(
-                    f"Can only provide keyword {parameter!r} for "
+                    f"Can't set keyword {parameter!r} for "
                     f"quantization algorithm {algorithm!r}"
                 )
 
-            q.set_parameter(parameter, quantization_nsb, copy=False)
+            q.set_parameter(parameter, quantization_nsb)
 
         self._set_component("quantize_on_write", q, copy=False)